requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.14.1
mysqlclient==2.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
requests==2.31.0
schedule==1.2.1
Flask==3.1.1